collect mail attachment related tests to same class
